combines filters and pinned statuses into one account context
